StdLib color conversions

+++ Assets/Scripts/Includes/ColorConversion.cginc
@@ -1,6 +1,6 @@
 ﻿#ifndef COLOR_CONVERSIONS
 #define COLOR_CONVERSIONS
-//Taken from the Pos Processing stack StdLib
+//Taken from the Post Processing stack StdLib
 // -----------------------------------------------------------------------------
 // Constants
 
@@ -49,70 +49,70 @@ float4 PositivePow(float4 base, float4 power)
 // Fast path ref: http://chilliant.blogspot.com.au/2012/08/srgb-approximations-for-hlsl.html?m=1
 //
 
-half SRGBToLinear(half c)
+float SRGBToLinear(float c)
 {
     #if USE_VERY_FAST_SRGB
     return c * c;
     #elif USE_FAST_SRGB
     return c * (c * (c * 0.305306011 + 0.682171111) + 0.012522878);
     #else
-    half linearRGBLo = c / 12.92;
-    half linearRGBHi = PositivePow((c + 0.055) / 1.055, 2.4);
-    half linearRGB = (c <= 0.04045) ? linearRGBLo : linearRGBHi;
+    float linearRGBLo = c / 12.92;
+    float linearRGBHi = PositivePow((c + 0.055) / 1.055, 2.4);
+    float linearRGB = (c <= 0.04045) ? linearRGBLo : linearRGBHi;
     return linearRGB;
     #endif
 }
 
-half3 SRGBToLinear(half3 c)
+float3 SRGBToLinear(float3 c)
 {
     #if USE_VERY_FAST_SRGB
     return c * c;
     #elif USE_FAST_SRGB
     return c * (c * (c * 0.305306011 + 0.682171111) + 0.012522878);
     #else
-    half3 linearRGBLo = c / 12.92;
-    half3 linearRGBHi = PositivePow((c + 0.055) / 1.055, half3(2.4, 2.4, 2.4));
-    half3 linearRGB = (c <= 0.04045) ? linearRGBLo : linearRGBHi;
+    float3 linearRGBLo = c / 12.92;
+    float3 linearRGBHi = PositivePow((c + 0.055) / 1.055, float3(2.4, 2.4, 2.4));
+    float3 linearRGB = (c <= 0.04045) ? linearRGBLo : linearRGBHi;
     return linearRGB;
     #endif
 }
 
-half4 SRGBToLinear(half4 c)
+float4 SRGBToLinear(float4 c)
 {
-    return half4(SRGBToLinear(c.rgb), c.a);
+    return float4(SRGBToLinear(c.rgb), c.a);
 }
 
-half LinearToSRGB(half c)
+float LinearToSRGB(float c)
 {
     #if USE_VERY_FAST_SRGB
     return sqrt(c);
     #elif USE_FAST_SRGB
     return max(1.055 * PositivePow(c, 0.416666667) - 0.055, 0.0);
     #else
-    half sRGBLo = c * 12.92;
-    half sRGBHi = (PositivePow(c, 1.0 / 2.4) * 1.055) - 0.055;
-    half sRGB = (c <= 0.0031308) ? sRGBLo : sRGBHi;
+    float sRGBLo = c * 12.92;
+    float sRGBHi = (PositivePow(c, 1.0 / 2.4) * 1.055) - 0.055;
+    float sRGB = (c <= 0.0031308) ? sRGBLo : sRGBHi;
     return sRGB;
     #endif
 }
 
-half3 LinearToSRGB(half3 c)
+float3 LinearToSRGB(float3 c)
 {
     #if USE_VERY_FAST_SRGB
     return sqrt(c);
     #elif USE_FAST_SRGB
     return max(1.055 * PositivePow(c, 0.416666667) - 0.055, 0.0);
     #else
-    half3 sRGBLo = c * 12.92;
-    half3 sRGBHi = (PositivePow(c, half3(1.0 / 2.4, 1.0 / 2.4, 1.0 / 2.4)) * 1.055) - 0.055;
-    half3 sRGB = (c <= 0.0031308) ? sRGBLo : sRGBHi;
+    float3 sRGBLo = c * 12.92;
+    float3 sRGBHi = (PositivePow(c, float3(1.0 / 2.4, 1.0 / 2.4, 1.0 / 2.4)) * 1.055) - 0.055;
+    float3 sRGB = (c <= 0.0031308) ? sRGBLo : sRGBHi;
     return sRGB;
     #endif
 }
 
-half4 LinearToSRGB(half4 c)
+float4 LinearToSRGB(float4 c)
 {
-    return half4(LinearToSRGB(c.rgb), c.a);
+    return float4(LinearToSRGB(c.rgb), c.a);
 }
 
 #endif
